quirks: Force X11 on Linux if certain symbols are in the process.
So if we see the process references GLEW or Nvidia Cg, we assume the
app will fail if it can't talk to glX.

This automates the process in many cases instead of having to list these
specific titles in the quirks table.

This patch in its initial form, and the idea, are from @smvc (thanks!).

Fixes #249.
Fixes #252.
Fixes #258.
Fixes #262.
Fixes #263.