video: If trying to create a < 8 bit surface, force it to 8 bits.
This gets around a specific game (rockdodger) that creates a 2-bit surface
just to convert it to display format.

This is likely to confuse any game that wants to directly manipulate pixels
and isn't getting the surface they expected, but until now those games were
simply failing here, so we'll revisit with a more-complex solution when the
need arises.

Fixes #260.