video: Fix apps that UpdateRect and never PumpEvents.
An app that _only_ calls SDL_UpdateRects (perhaps to draw just a piece of
animation that doesn't cover the screen but while not accepting user input)
will now check inside SDL_UpdateRects itself to decide if enough time has
gone by to warrant a present.

Also, if SDL_Delay is called and a present is pending at all, just do it,
as a signal from the app that the frame is complete. This catches things
that update a rectangle and then delay for the user to view it.

Fixes #253.