quirks: Force X11 if plib (specifically libplibssg) is detected
crrcsim, a model aeroplane simulator, is linked to both SDL 1.2 and
plib. I'm not entirely sure why, since plib seems to be an (unmaintained)
alternative to SDL that did many of the same things.

Because plib has a C++ ABI but dlsym() works with C-level ABI, we need
to use a C++ mangled name for the symbol we look for.

Resolves: https://github.com/libsdl-org/sdl12-compat/issues/252
Signed-off-by: Simon McVittie <smcv@collabora.com>