audio: Add a quirk to mimick SDL 1.2's audio converters more closely.
In practice, you _don't_ want this; the resampler is extremely low
quality and will overflow buffers if the frequency conversion isn't a
multiple of two (just like SDL 1.2), but it works if the app allocates
its conversion buffer incorrectly, and it's possible that some app
relies on the quirky conversion in some form, too.

Fixes #232.