SDL_config: Define HAVE_MALLOC_H on glibc-based platforms
Version 3.7.4 of the Freespace 2 Source Code Project is known to rely
on SDL having checked for this, and won't see a declaration of
malloc_usable_size() if we don't.

This is a workaround for source compatibility with classic SDL 1.2:
if a game wants to use functions from a particular system header file,
it should be that game's responsibility to check for the header file
itself.

Bug-Debian: https://bugs.debian.org/1012232#58
Signed-off-by: Simon McVittie <smcv@debian.org>