Hint: Add SDL12COMPAT_WINDOWED_MODE_LIST hint, to make SDL_ListModes() work in windowed mode.
Add a new boolean hint, which makes SDL_ListModes() always return a mode list,
even if the flags passed in are for windowed mode. By default, we return -1
in this case, to indicate that any mode is supported.

However, some older titles pass 0 for flags here even if they're intending to
run in fullscreen, and if -1 is returned, attempt to populate their own mode
list (and select a mode) themselves. For example, Civilization: Call To Power
will add the current desktop mode (if it can glean it from X11, which doesn't
work on Wayland, and gives an unusuable combined size for multi-monitor setups
even on X11), along with the hardcoded list of 1024×768, 800×600, and 640×480).

By always returning sdl12-compat's list of modes (some of which may be "fake"),
we get a much more modern list of supported modes, which allows the game to
comfortably run at higher resolutions.